# docs/swagger.yaml
basePath: /api
definitions:
...
  handler.ErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      message:
        example: invalid input
        type: string
      status:
        example: fail
        type: string
    type: object
...
  handler.LoginResponse:
    description: Токены и информация о пользователе после входа
    properties:
      is_moderator:
        example: false
        type: boolean
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      username:
        example: user1
        type: string
    type: object
...
  handler.RefreshRequest:
    description: Refresh токен для получения новой пары токенов
    properties:
      refresh_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    required:
    - refresh_token
    type: object
  handler.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      status:
        example: ok
        type: string
    type: object
  handler.UserLoginRequest:
    description: Данные для входа в систему
    properties:
      password:
        example: password
        type: string
      username:
        example: user1
        type: string
    required:
    - password
    - username
    type: object
  handler.UserRegisterRequest:
    description: Данные для регистрации нового пользователя
    properties:
      password:
        example: password123
        type: string
      username:
        example: newuser
        type: string
    required:
    - password
    - username
    type: object
  handler.UserResponse:
    description: Информация о пользователе
    properties:
      is_moderator:
        example: false
        type: boolean
      username:
        example: user1
        type: string
    type: object
  handler.UserUpdateRequest:
    description: Данные для обновления профиля пользователя
    properties:
      password:
        example: newpassword123
        type: string
      username:
        example: newusername
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@cavi.local
    name: API Support
  description: |-
    API для расчёта индекса CAVI (Cardio-Ankle Vascular Index).
    Система позволяет управлять группами пациентов, создавать заявки на расчёт,
    добавлять группы в заявки и проводить модерацию.
  title: CAVI Calculator API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Аутентифицирует пользователя и возвращает JWT токены (access +
        refresh).
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Неверные учётные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Вход
      tags:
      - users
  /users/logout:
    post:
      description: Добавляет access токен в blacklist Redis и удаляет refresh токен.
      produces:
      - application/json
      responses:
        "200":
          description: Успешный выход
          schema:
            $ref: '#/definitions/handler.SuccessResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Выход
      tags:
      - users
  /users/me:
    get:
      description: Возвращает информацию о текущем авторизованном пользователе.
      produces:
      - application/json
      responses:
        "200":
          description: Данные пользователя
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Текущий пользователь
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет username и/или пароль текущего пользователя.
      parameters:
      - description: Данные для обновления
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль обновлён
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Неверные данные или username занят
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Требуется аутентификация
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить профиль
      tags:
      - users
  /users/refresh:
    post:
      consumes:
      - application/json
      description: Обновляет access и refresh токены по валидному refresh токену.
      parameters:
      - description: Refresh токен
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Новые токены
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Неверные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Невалидный или истёкший refresh токен
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Обновить токены
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя в системе. Пароль хешируется с помощью
        bcrypt.
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/handler.UserResponse'
        "400":
          description: Неверные данные или пользователь существует
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      summary: Регистрация
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'JWT токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"